berhasil seeder baru

+++ app/Http/Controllers/Backsite/SpecialistController.php
@@ -13,7 +13,7 @@ use App\Http\Specialists\Request\StoreSpecialistRequest;
 use App\Http\Specialists\Request\UpdateSpecialistRequest;
 
 //user everything
-//use Gate;
+use Gate;
 use Auth;
 
 //model here
@@ -36,6 +36,8 @@ class SpecialistController extends Controller
 
     public function index()
     {
+        abort_if(Gate::denies('specialist_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $specialist = Specialist::orderBy('created_at', 'desc')->get();
 
         return view('pages.backsite.master-data.specialist.index');
@@ -78,6 +80,8 @@ class SpecialistController extends Controller
      */
     public function show(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('pages.backsite.master-data.specialist.show', compact('specialist'));
     }
 
@@ -89,6 +93,8 @@ class SpecialistController extends Controller
      */
     public function edit(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('pages.backsite.master-data.specialist.edit', compact('specialist'));
     }
 
@@ -120,6 +126,8 @@ class SpecialistController extends Controller
      */
     public function destroy(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $specialist->delete();
 
         alert()->success('Success Message', 'Successfully deleted specialist');
